LPS-94547 adapt
