--- modules/apps/headless/headless-form/headless-form-test/src/testIntegration/java/com/liferay/headless/form/resource/v1_0/test/FormRecordFormResourceTest.java
@@ -1,29 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.headless.form.resource.v1_0.test;
-
-import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-
-import org.junit.Ignore;
-import org.junit.runner.RunWith;
-
-/**
- * @author Javier Gamarra
- */
-@Ignore
-@RunWith(Arquillian.class)
-public class FormRecordFormResourceTest
-	extends BaseFormRecordFormResourceTestCase {
-}
